JWT Error

+++ services/posts.js
@@ -1,8 +1,9 @@
 const MongoLib = require('../lib/mongo');
+const { unauthorized } = require('@hapi/boom');
 
 class PostsService {
   constructor() {
-    this.collection = 'pots';
+    this.collection = 'posts';
     this.DB = new MongoLib();
   }
 
@@ -16,11 +17,19 @@ class PostsService {
     return createdPost;
   }
 
-  async updatePost(id, data) {
+  async updatePost(id, data, user) {
+    const post = await this.getPost(id);
+    if (post.user !== user) {
+      return unauthorized();
+    }
     const updatedPost = await this.DB.update(this.collection, id, data);
     return updatedPost;
   }
-  async deletePost(id) {
+  async deletePost(id, user) {
+    const post = await this.getPost(id);
+    if (post.user !== user) {
+      return unauthorized();
+    }
     const deletedPost = await this.DB.delete(this.collection, id);
     return deletedPost;
   }
